Use tree representation for graph layout
Graph representation doesn't look correct in some cases, so use tree
representation instead which appears to be proper. I didn't remove graph API
endpoint in this PR as I wasn't sure if there're any plans for it, but it's
unused now.

I added some tests in the frontend after the refactoring and not before just to
avoid extra work of adjusting the [invalid when using graph endpoint] input
test data, so basically rewriting tests that weren't there to begin with.
Hopefully that's okay.