Fix tree representation for scripted parallel stages
Use node parent ID for synthetic parallel node so that it has a valid ID which
we will be able to get the node for from execution in `PipelineGraphApi`'s
`createTree()` later on.  In an unlikely event of this failing, still fallback
to ID generation based on appending "-parallel-synthetic" to the ID of first
parallel node.