Fix caching of incomplete tree/allSteps
There was a race condition between the computation of the tree/allSteps
and looking up the completed state of the run.
- time 00:00 the run is pending
- time 00:01 start computing the tree/allSteps
- time 00:02 the run completes
- time 00:03 computing the tree/allSteps finishes
- time 00:04 call "setCache"

Previously, "setCache" would see the run as completed and emit caching
headers. Now, the completed state is collected before computing the tree
or allSteps. The tree/allSteps might represent the completed run, but
the frontend will need to poll it one more time to make sure.