Fix scroll conflict between stickToBottom and selecting a log line
When scrolling to a selected line we may scroll close to the bottom of
the viewport. The smooth scrolling will emit multiple scroll events.
Do not take any of them as user initiated or intent for staying at the
bottom of the log lines.

Effectively, abort scrolling to the bottom as soon as a log line was
selected.