Optimize ScrollToTopBottom
Avoid listening to ALL DOM changes. The attributes of "running" icons
are constantly changing and triggering events.
Instead, use an IntersectionObserver for checking the visibility of the
top row of UI elements below the Jenkins header (isAtTop) and a dummy
element at the bottom of the UI (isAtBottom).

Additionally, switch the scrollToBottom to scroll to the dummy element
at the bottom. This will account for any pending DOM changes and line up
the bottom of the logs with the bottom of the view -- rather than
scrolling to the Jenkins version.